New TaskEvent class, split objs file into folder
