+++ data/backup/worker.csv
@@ -1,0 +1,9 @@
+,worker,action,stimulus_id,time
+0,tcp://10.201.0.212:38553,add,,1713455674.0437584
+1,tcp://10.201.0.212:34699,add,,1713455674.579557
+2,tcp://10.201.0.212:45449,add,,1713455674.5826244
+3,tcp://10.201.0.212:38577,add,,1713455674.583946
+4,tcp://10.201.0.229:37485,add,,1713455675.4265223
+5,tcp://10.201.0.229:41793,add,,1713455675.427957
+6,tcp://10.201.0.229:44955,add,,1713455675.4291391
+7,tcp://10.201.0.229:45945,add,,1713455675.4302516
