+++ ex_run.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+python ./src/dask_capture.py -o ./data/compiled_tasks.txt --debug ./data
